+++ CI/SCRIPTS/test_types.py
@@ -19,7 +19,8 @@ def test_is_iterable():
     assert is_iterable([1, 2, 3])
     assert is_iterable({1, 2, 3})
     assert is_iterable(np.array([1, 2, 3]))
-    assert is_iterable("1, 2, 3")
+    assert not is_iterable("1, 2, 3")
+    assert is_iterable("1, 2, 3", str_allowed=True)
     assert not is_iterable(1)
     assert not is_iterable(AnyPath("1, 2, 3"))
 
@@ -27,11 +28,14 @@ def test_is_iterable():
 def test_make_iterable():
     """Test make_iterable"""
 
-    def assert_mi(val, is_it=True):
+    def assert_mi(val, is_it=True, **kwargs):
         if not is_it:
             val = [val]
 
-        comp = val == make_iterable(val)
+        if val is None:
+            comp = True
+        else:
+            comp = val == make_iterable(val, **kwargs)
         try:
             assert comp
         except ValueError:
@@ -41,6 +45,8 @@ def test_make_iterable():
     assert_mi([1, 2, 3])
     assert_mi({1, 2, 3})
     assert_mi(np.array([1, 2, 3]))
-    assert_mi("1, 2, 3")
+    assert_mi("1, 2, 3", str_allowed=True)
+    assert_mi("1, 2, 3", str_allowed=False, is_it=False)
     assert_mi(1, is_it=False)
-    assert_mi(AnyPath("1, 2, 3"), is_it=False)
+    assert_mi(None, convert_none=True)
+    assert_mi(None, convert_none=False, is_it=False)